factor out three hashmaps in Config into a new struct for better encapsulation
